Доработать функциональность Пользователи и Категории
