+++ main-service/src/main/java/ru/practicum/ewm/MainApplication.java
@@ -1,0 +1,14 @@
+package ru.practicum.ewm;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+
+
+@SpringBootApplication
+public class MainApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(MainApplication.class, args);
+    }
+}
