+++ main-service/src/main/java/ru/practicum/ewm/category/dto/NewCategoryDto.java
@@ -1,0 +1,17 @@
+package ru.practicum.ewm.category.dto;
+
+import jakarta.validation.constraints.NotBlank;
+import jakarta.validation.constraints.Size;
+import lombok.*;
+import lombok.experimental.FieldDefaults;
+
+@Getter
+@FieldDefaults(level = AccessLevel.PRIVATE)
+@NoArgsConstructor
+@AllArgsConstructor
+@Builder
+public class NewCategoryDto {
+    @NotBlank
+    @Size(min = 1, max = 50)
+    String name;
+}
