+++ main-service/src/main/java/ru/practicum/ewm/category/mapper/CategoryMapper.java
@@ -1,0 +1,22 @@
+package ru.practicum.ewm.category.mapper;
+
+import lombok.experimental.UtilityClass;
+import ru.practicum.ewm.category.dto.CategoryDto;
+import ru.practicum.ewm.category.dto.NewCategoryDto;
+import ru.practicum.ewm.category.model.Category;
+
+@UtilityClass
+public class CategoryMapper {
+    public Category toCategory(NewCategoryDto newCategoryDto) {
+        return Category.builder()
+                .name(newCategoryDto.getName())
+                .build();
+    }
+
+    public CategoryDto toCategoryDto(Category category) {
+        return CategoryDto.builder()
+                .id(category.getId())
+                .name(category.getName())
+                .build();
+    }
+}
