+++ main-service/src/main/java/ru/practicum/ewm/category/service/CategoryService.java
@@ -1,0 +1,18 @@
+package ru.practicum.ewm.category.service;
+
+import ru.practicum.ewm.category.dto.CategoryDto;
+import ru.practicum.ewm.category.dto.NewCategoryDto;
+
+import java.util.List;
+
+public interface CategoryService {
+    CategoryDto createCategory(NewCategoryDto newCategoryDto);
+
+    void deleteCategory(Long catId);
+
+    CategoryDto updateCategory(CategoryDto categoryDto, Long catId);
+
+    List<CategoryDto> getCategories(Integer from, Integer size);
+
+    CategoryDto getCategoryById(Long catId);
+}
