+++ main-service/src/main/java/ru/practicum/ewm/category/service/CategoryServiceImpl.java
@@ -1,0 +1,77 @@
+package ru.practicum.ewm.category.service;
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.data.domain.PageRequest;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+import ru.practicum.ewm.category.dto.CategoryDto;
+import ru.practicum.ewm.category.dto.NewCategoryDto;
+import ru.practicum.ewm.category.mapper.CategoryMapper;
+import ru.practicum.ewm.category.model.Category;
+import ru.practicum.ewm.category.repository.CategoryRepository;
+import ru.practicum.ewm.exception.DuplicateException;
+import ru.practicum.ewm.exception.NotFoundException;
+
+import java.util.List;
+
+
+@Service
+@RequiredArgsConstructor
+@Transactional(readOnly = true)
+public class CategoryServiceImpl implements CategoryService {
+
+    private final CategoryRepository categoryRepository;
+
+    @Transactional
+    @Override
+    public CategoryDto createCategory(NewCategoryDto newCategoryDto) {
+        if (categoryRepository.existsByName(newCategoryDto.getName())) {
+            throw new DuplicateException("Категория с таким именем уже существует");
+        }
+        Category category = categoryRepository.save(CategoryMapper.toCategory(newCategoryDto));
+        return CategoryMapper.toCategoryDto(category);
+    }
+
+    @Transactional
+    @Override
+    public void deleteCategory(Long catId) {
+        if (!categoryRepository.existsById(catId)) {
+            throw new NotFoundException("Категория не найдена");
+        }
+        categoryRepository.deleteById(catId);
+    }
+
+    @Transactional
+    @Override
+    public CategoryDto updateCategory(CategoryDto categoryDto, Long catId) {
+        Category category = checkCategory(catId);
+
+        if (!category.getName().equals(categoryDto.getName())) {
+            if (categoryRepository.existsByName(categoryDto.getName())) {
+                throw new DuplicateException("Категория с таким именем уже существует");
+            }
+            category.setName(categoryDto.getName());
+        }
+
+        return CategoryMapper.toCategoryDto(category);
+    }
+
+    @Override
+    public List<CategoryDto> getCategories(Integer from, Integer size) {
+        PageRequest pageRequest = PageRequest.of(from / size, size);
+
+        return categoryRepository.findAll(pageRequest).stream()
+                .map(CategoryMapper::toCategoryDto)
+                .toList();
+    }
+
+    @Override
+    public CategoryDto getCategoryById(Long catId) {
+        return CategoryMapper.toCategoryDto(checkCategory(catId));
+    }
+
+    private Category checkCategory(Long catId) {
+        return categoryRepository.findById(catId)
+                .orElseThrow(() -> new NotFoundException("Категория не найдена или недоступна"));
+    }
+}
