+++ main-service/src/main/java/ru/practicum/ewm/error/ErrorResponse.java
@@ -1,0 +1,3 @@
+package ru.practicum.ewm.error;
+public record ErrorResponse(String error) {
+}
