+++ main-service/src/main/java/ru/practicum/ewm/exception/DuplicateException.java
@@ -1,0 +1,7 @@
+package ru.practicum.ewm.exception;
+
+public class DuplicateException extends RuntimeException {
+    public DuplicateException(String message) {
+        super(message);
+    }
+}
