+++ main-service/src/main/java/ru/practicum/ewm/user/dto/UserDto.java
@@ -1,0 +1,16 @@
+package ru.practicum.ewm.user.dto;
+
+import lombok.*;
+import lombok.experimental.FieldDefaults;
+
+
+@Getter
+@FieldDefaults(level = AccessLevel.PRIVATE)
+@NoArgsConstructor
+@AllArgsConstructor
+@Builder
+public class UserDto {
+    Long id;
+    String email;
+    String name;
+}
