+++ main-service/src/main/java/ru/practicum/ewm/user/mapper/UserMapper.java
@@ -1,0 +1,32 @@
+package ru.practicum.ewm.user.mapper;
+
+import lombok.experimental.UtilityClass;
+import ru.practicum.ewm.user.dto.NewUserRequest;
+import ru.practicum.ewm.user.dto.UserDto;
+import ru.practicum.ewm.user.dto.UserShortDto;
+import ru.practicum.ewm.user.model.User;
+
+@UtilityClass
+public class UserMapper {
+    public User toUser(NewUserRequest newUserRequest) {
+        return User.builder()
+                .email(newUserRequest.getEmail())
+                .name(newUserRequest.getName())
+                .build();
+    }
+
+    public UserDto toUserDto(User user) {
+        return UserDto.builder()
+                .email(user.getEmail())
+                .id(user.getId())
+                .name(user.getName())
+                .build();
+    }
+
+    public UserShortDto toUserShortDto(User user) {
+        return UserShortDto.builder()
+                .id(user.getId())
+                .name(user.getName())
+                .build();
+    }
+}
