+++ main-service/src/main/java/ru/practicum/ewm/user/model/User.java
@@ -1,0 +1,28 @@
+package ru.practicum.ewm.user.model;
+
+import jakarta.persistence.*;
+import lombok.*;
+import lombok.experimental.FieldDefaults;
+
+
+@Entity
+@Table(name = "users")
+@ToString
+@Getter
+@Setter
+@AllArgsConstructor
+@NoArgsConstructor
+@FieldDefaults(level = AccessLevel.PRIVATE)
+@Builder(toBuilder = true)
+public class User {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    @Column(name = "user_id")
+    Long id;
+
+    @Column(unique = true)
+    String email;
+
+    String name;
+}
