+++ main-service/src/main/java/ru/practicum/ewm/user/repository/UserRepository.java
@@ -1,0 +1,15 @@
+package ru.practicum.ewm.user.repository;
+
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.PageRequest;
+import org.springframework.data.jpa.repository.JpaRepository;
+import ru.practicum.ewm.user.model.User;
+
+import java.util.List;
+
+public interface UserRepository extends JpaRepository<User, Long> {
+
+    Page<User> findAllByIdIn(List<Long> ids, PageRequest pageRequest);
+
+    Boolean existsByEmail(String email);
+}
