+++ main-service/src/main/java/ru/practicum/ewm/user/service/UserServiceImpl.java
@@ -1,0 +1,57 @@
+package ru.practicum.ewm.user.service;
+
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.data.domain.PageRequest;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+import ru.practicum.ewm.exception.DuplicateException;
+import ru.practicum.ewm.exception.NotFoundException;
+import ru.practicum.ewm.user.dto.NewUserRequest;
+import ru.practicum.ewm.user.dto.UserDto;
+import ru.practicum.ewm.user.mapper.UserMapper;
+import ru.practicum.ewm.user.repository.UserRepository;
+
+import java.util.List;
+
+@Service
+@RequiredArgsConstructor
+@Transactional(readOnly = true)
+@Slf4j
+public class UserServiceImpl implements UserService {
+
+    private final UserRepository userRepository;
+
+    @Override
+    public List<UserDto> getAllUsers(List<Long> ids, Integer from, Integer size) {
+        PageRequest pageRequest = PageRequest.of(from / size, size);
+
+        if (ids.isEmpty()) {
+            return userRepository.findAll(pageRequest).getContent().stream()
+                    .map(UserMapper::toUserDto)
+                    .toList();
+        } else {
+            return userRepository.findAllByIdIn(ids, pageRequest).getContent().stream()
+                    .map(UserMapper::toUserDto)
+                    .toList();
+        }
+    }
+
+    @Transactional
+    @Override
+    public UserDto saveUser(NewUserRequest newUserRequest) {
+        if (userRepository.existsByEmail(newUserRequest.getEmail())) {
+            throw new DuplicateException("Пользователь с таким email уже существует");
+        }
+        return UserMapper.toUserDto(userRepository.save(UserMapper.toUser(newUserRequest)));
+    }
+
+    @Transactional
+    @Override
+    public void deleteUser(Long id) {
+        if (!userRepository.existsById(id)) {
+            throw new NotFoundException("Пользователь не найден");
+        }
+        userRepository.deleteById(id);
+    }
+}
