--- main-service/src/main/java/ru/practicum/ewmservice/MainApplication.java
@@ -1,14 +1,0 @@
-package ru.practicum.ewmservice;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-
-
-@SpringBootApplication
-public class MainApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(MainApplication.class, args);
-    }
-}
