--- main-service/src/main/java/ru/practicum/ewmservice/category/controller/AdminCategoryController.java
@@ -1,40 +1,0 @@
-package ru.practicum.ewmservice.category.controller;
-
-import jakarta.validation.Valid;
-import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.http.HttpStatus;
-import org.springframework.web.bind.annotation.*;
-import ru.practicum.ewmservice.category.dto.CategoryDto;
-import ru.practicum.ewmservice.category.dto.NewCategoryDto;
-import ru.practicum.ewmservice.category.service.CategoryService;
-
-@RestController
-@RequiredArgsConstructor
-@RequestMapping("admin/categories")
-@Slf4j
-public class AdminCategoryController {
-
-    private final CategoryService categoryService;
-
-    @PostMapping
-    @ResponseStatus(HttpStatus.CREATED)
-    public CategoryDto createCategory(@RequestBody @Valid NewCategoryDto newCategoryDto) {
-        log.info("Пришел запрос на создание категории.");
-        return categoryService.createCategory(newCategoryDto);
-    }
-
-    @DeleteMapping("/{catId}")
-    @ResponseStatus(HttpStatus.NO_CONTENT)
-    public void deleteCategory(@PathVariable Long catId) {
-        log.info("Пришел запрос на удаление категории.");
-        categoryService.deleteCategory(catId);
-    }
-
-    @PatchMapping("/{catId}")
-    public CategoryDto updateCategory(@RequestBody @Valid CategoryDto categoryDto,
-                                      @PathVariable Long catId) {
-        log.info("Пришел запрос на обновление категории.");
-        return categoryService.updateCategory(categoryDto, catId);
-    }
-}
