--- main-service/src/main/java/ru/practicum/ewmservice/category/controller/PublicCategoryController.java
@@ -1,34 +1,0 @@
-package ru.practicum.ewmservice.category.controller;
-
-import jakarta.validation.constraints.Positive;
-import jakarta.validation.constraints.PositiveOrZero;
-import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.web.bind.annotation.*;
-import ru.practicum.ewmservice.category.dto.CategoryDto;
-import ru.practicum.ewmservice.category.service.CategoryService;
-
-import java.util.List;
-
-@RestController
-@RequiredArgsConstructor
-@RequestMapping("/categories")
-@Slf4j
-public class PublicCategoryController {
-
-    private final CategoryService categoryService;
-
-    @GetMapping
-    public List<CategoryDto> getCategories(
-            @PositiveOrZero @RequestParam(name = "from", defaultValue = "0")   Integer from,
-            @Positive @RequestParam(name = "size", defaultValue = "10") Integer size) {
-        log.info("Пришел запрос на получение списка категорий");
-        return categoryService.getCategories(from,size);
-    }
-
-    @GetMapping("/{catId}")
-    public CategoryDto getCategoryById(@PathVariable Long catId) {
-        log.info("Пришел запрос на получение категории по id.");
-        return categoryService.getCategoryById(catId);
-    }
-}
