--- main-service/src/main/java/ru/practicum/ewmservice/category/dto/CategoryDto.java
@@ -1,18 +1,0 @@
-package ru.practicum.ewmservice.category.dto;
-
-import jakarta.validation.constraints.NotBlank;
-import jakarta.validation.constraints.Size;
-import lombok.*;
-import lombok.experimental.FieldDefaults;
-
-@Getter
-@FieldDefaults(level = AccessLevel.PRIVATE)
-@NoArgsConstructor
-@AllArgsConstructor
-@Builder
-public class CategoryDto {
-    private Long id;
-    @NotBlank
-    @Size(min = 1,max = 50)
-    private String name;
-}
