--- main-service/src/main/java/ru/practicum/ewmservice/category/mapper/CategoryMapper.java
@@ -1,22 +1,0 @@
-package ru.practicum.ewmservice.category.mapper;
-
-import lombok.experimental.UtilityClass;
-import ru.practicum.ewmservice.category.dto.CategoryDto;
-import ru.practicum.ewmservice.category.dto.NewCategoryDto;
-import ru.practicum.ewmservice.category.model.Category;
-
-@UtilityClass
-public class CategoryMapper {
-    public Category toCategory (NewCategoryDto newCategoryDto) {
-        return Category.builder()
-                .name(newCategoryDto.getName())
-                .build();
-    }
-
-    public CategoryDto toCategoryDto (Category category) {
-        return CategoryDto.builder()
-                .id(category.getId())
-                .name(category.getName())
-                .build();
-    }
-}
