--- main-service/src/main/java/ru/practicum/ewmservice/category/model/Category.java
@@ -1,24 +1,0 @@
-package ru.practicum.ewmservice.category.model;
-
-
-import jakarta.persistence.*;
-import lombok.*;
-import lombok.experimental.FieldDefaults;
-
-@Entity
-@Table(name = "categories")
-@ToString
-@Getter
-@Setter
-@AllArgsConstructor
-@NoArgsConstructor
-@FieldDefaults(level = AccessLevel.PRIVATE)
-@Builder(toBuilder = true)
-public class Category {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "category_id")
-    Long id;
-    @Column(unique = true)
-    String name;
-}
