--- main-service/src/main/java/ru/practicum/ewmservice/category/service/CategoryService.java
@@ -1,18 +1,0 @@
-package ru.practicum.ewmservice.category.service;
-
-import ru.practicum.ewmservice.category.dto.CategoryDto;
-import ru.practicum.ewmservice.category.dto.NewCategoryDto;
-
-import java.util.List;
-
-public interface CategoryService {
-    CategoryDto createCategory(NewCategoryDto newCategoryDto);
-
-    void deleteCategory(Long catId);
-
-    CategoryDto updateCategory(CategoryDto categoryDto, Long catId);
-
-    List<CategoryDto> getCategories(Integer from, Integer size);
-
-    CategoryDto getCategoryById(Long catId);
-}
