--- main-service/src/main/java/ru/practicum/ewmservice/category/service/CategoryServiceImpl.java
@@ -1,63 +1,0 @@
-package ru.practicum.ewmservice.category.service;
-
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
-import ru.practicum.ewmservice.category.dto.CategoryDto;
-import ru.practicum.ewmservice.category.dto.NewCategoryDto;
-import ru.practicum.ewmservice.category.mapper.CategoryMapper;
-import ru.practicum.ewmservice.category.model.Category;
-import ru.practicum.ewmservice.category.repository.CategoryRepository;
-import ru.practicum.ewmservice.exception.NotFoundException;
-
-import java.util.List;
-
-
-@Service
-@RequiredArgsConstructor
-@Transactional(readOnly = true)
-public class CategoryServiceImpl implements CategoryService {
-
-    private final CategoryRepository categoryRepository;
-
-    @Transactional
-    @Override
-    public CategoryDto createCategory(NewCategoryDto newCategoryDto) {
-        Category category = categoryRepository.save(CategoryMapper.toCategory(newCategoryDto));
-        return CategoryMapper.toCategoryDto(category);
-    }
-
-    @Transactional
-    @Override
-    public void deleteCategory(Long catId) {
-        checkCategory(catId);
-        categoryRepository.deleteById(catId);
-    }
-
-    @Transactional
-    @Override
-    public CategoryDto updateCategory(CategoryDto categoryDto, Long catId) {
-        Category category = checkCategory(catId);
-        category.setName(categoryDto.getName());
-        return CategoryMapper.toCategoryDto(category);
-    }
-
-    @Override
-    public List<CategoryDto> getCategories(Integer from, Integer size) {
-        return categoryRepository.findAll().stream()
-                .skip(from)
-                .limit(size)
-                .map(CategoryMapper::toCategoryDto)
-                .toList();
-    }
-
-    @Override
-    public CategoryDto getCategoryById(Long catId) {
-        return CategoryMapper.toCategoryDto(checkCategory(catId));
-    }
-
-    private Category checkCategory(Long catId) {
-        return categoryRepository.findById(catId)
-                .orElseThrow(() -> new NotFoundException("Категория не найдена или недоступна"));
-    }
-}
