--- main-service/src/main/java/ru/practicum/ewmservice/error/ErrorResponse.java
@@ -1,3 +1,0 @@
-package ru.practicum.ewmservice.error;
-public record ErrorResponse(String error) {
-}
