--- main-service/src/main/java/ru/practicum/ewmservice/user/controller/UserController.java
@@ -1,45 +1,0 @@
-package ru.practicum.ewmservice.user.controller;
-
-
-import jakarta.validation.Valid;
-import jakarta.validation.constraints.Positive;
-import jakarta.validation.constraints.PositiveOrZero;
-import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.http.HttpStatus;
-import org.springframework.web.bind.annotation.*;
-import ru.practicum.ewmservice.user.service.UserService;
-import ru.practicum.ewmservice.user.dto.NewUserRequest;
-import ru.practicum.ewmservice.user.dto.UserDto;
-
-import java.util.List;
-
-@RestController
-@RequiredArgsConstructor
-@RequestMapping("/admin/users")
-@Slf4j
-public class UserController {
-    private final UserService userService;
-
-    @GetMapping()
-    public List<UserDto> getAllUsers(
-            @RequestParam(name = "ids", defaultValue = "") List<Long> ids,
-            @PositiveOrZero @RequestParam(name = "from", defaultValue = "0")  Integer from,
-            @Positive @RequestParam(name = "size", defaultValue = "10")  Integer size) {
-        log.info("Пришел запрос на получение списка пользователей");
-        return userService.getAllUsers(ids, from, size);
-    }
-
-    @PostMapping
-    @ResponseStatus(HttpStatus.CREATED)
-    public UserDto saveUser(@RequestBody @Valid NewUserRequest newUserRequest) {
-        log.info("Пришел запрос на добавление пользователя");
-        return userService.saveUser(newUserRequest);
-    }
-
-    @DeleteMapping("/{userId}")
-    @ResponseStatus(HttpStatus.NO_CONTENT)
-    public void deleteUser(@PathVariable("userId") Long userId) {
-        userService.deleteUser(userId);
-    }
-}
