--- main-service/src/main/java/ru/practicum/ewmservice/user/dto/UserShortDto.java
@@ -1,18 +1,0 @@
-package ru.practicum.ewmservice.user.dto;
-
-
-import jakarta.validation.constraints.NotBlank;
-import lombok.*;
-import lombok.experimental.FieldDefaults;
-
-
-@Getter
-@FieldDefaults(level = AccessLevel.PRIVATE)
-@NoArgsConstructor
-@AllArgsConstructor
-@Builder
-public class UserShortDto {
-    Long id;
-    @NotBlank
-    String name;
-}
