--- main-service/src/main/java/ru/practicum/ewmservice/user/mapper/UserMapper.java
@@ -1,32 +1,0 @@
-package ru.practicum.ewmservice.user.mapper;
-
-import lombok.experimental.UtilityClass;
-import ru.practicum.ewmservice.user.dto.NewUserRequest;
-import ru.practicum.ewmservice.user.dto.UserDto;
-import ru.practicum.ewmservice.user.dto.UserShortDto;
-import ru.practicum.ewmservice.user.model.User;
-
-@UtilityClass
-public class UserMapper {
-    public User toUser(NewUserRequest newUserRequest) {
-        return User.builder()
-                .email(newUserRequest.getEmail())
-                .name(newUserRequest.getName())
-                .build();
-    }
-
-    public UserDto toUserDto(User user) {
-        return UserDto.builder()
-                .email(user.getEmail())
-                .id(user.getId())
-                .name(user.getName())
-                .build();
-    }
-
-    public UserShortDto toUserShortDto(User user) {
-        return UserShortDto.builder()
-                .id(user.getId())
-                .name(user.getName())
-                .build();
-    }
-}
