--- main-service/src/main/java/ru/practicum/ewmservice/user/model/User.java
@@ -1,33 +1,0 @@
-package ru.practicum.ewmservice.user.model;
-
-import jakarta.persistence.*;
-import jakarta.validation.constraints.Email;
-import jakarta.validation.constraints.NotBlank;
-import lombok.*;
-import lombok.experimental.FieldDefaults;
-
-
-@Entity
-@Table(name = "users")
-@ToString
-@Getter
-@Setter
-@AllArgsConstructor
-@NoArgsConstructor
-@FieldDefaults(level = AccessLevel.PRIVATE)
-@Builder(toBuilder = true)
-public class User {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "user_id")
-    Long id;
-
-    @Email
-    @NotBlank
-    String email;
-
-    @NotBlank
-    String name;
-
-}
