--- main-service/src/main/java/ru/practicum/ewmservice/user/repository/UserRepository.java
@@ -1,7 +1,0 @@
-package ru.practicum.ewmservice.user.repository;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import ru.practicum.ewmservice.user.model.User;
-
-public interface UserRepository extends JpaRepository<User, Long> {
-}
