--- main-service/src/main/java/ru/practicum/ewmservice/user/service/UserService.java
@@ -1,15 +1,0 @@
-package ru.practicum.ewmservice.user.service;
-
-import ru.practicum.ewmservice.user.dto.NewUserRequest;
-import ru.practicum.ewmservice.user.dto.UserDto;
-
-import java.util.List;
-
-public interface UserService {
-
-    List<UserDto> getAllUsers(List<Long> ids, Integer from, Integer size);
-
-    UserDto saveUser(NewUserRequest newUserRequest);
-
-    void deleteUser(Long id);
-}
