--- main-service/src/main/java/ru/practicum/ewmservice/user/service/UserServiceImpl.java
@@ -1,54 +1,0 @@
-package ru.practicum.ewmservice.user.service;
-
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
-import ru.practicum.ewmservice.user.mapper.UserMapper;
-import ru.practicum.ewmservice.user.dto.NewUserRequest;
-import ru.practicum.ewmservice.user.dto.UserDto;
-import ru.practicum.ewmservice.user.model.User;
-import ru.practicum.ewmservice.user.repository.UserRepository;
-
-
-import java.util.List;
-
-@Service
-@RequiredArgsConstructor
-@Transactional(readOnly = true)
-public class UserServiceImpl implements UserService {
-
-    private final UserRepository userRepository;
-
-    @Override
-    public List<UserDto> getAllUsers(List<Long> ids, Integer from, Integer size) {
-        if(ids.isEmpty()) {
-            return userRepository.findAll()
-                    .stream()
-                    .skip(from)
-                    .limit(size)
-                    .map(UserMapper::toUserDto)
-                    .toList();
-        }
-        else {
-            return userRepository.findAllById(ids)
-                    .stream()
-                    .skip(from)
-                    .limit(size)
-                    .map(UserMapper::toUserDto)
-                    .toList();
-        }
-    }
-
-    @Transactional
-    @Override
-    public UserDto saveUser(NewUserRequest newUserRequest) {
-        User user = userRepository.save(UserMapper.toUser(newUserRequest));
-        return UserMapper.toUserDto(user);
-    }
-
-    @Transactional
-    @Override
-    public void deleteUser(Long id) {
-        userRepository.deleteById(id);
-    }
-}
