+++ main-service/src/main/resources/application.properties
@@ -4,8 +4,7 @@ spring.jpa.hibernate.ddl-auto=create
 spring.jpa.properties.hibernate.format_sql=true
 spring.sql.init.mode=always
 
-
 spring.datasource.driverClassName=org.postgresql.Driver
-spring.datasource.url=jdbc:postgresql://localhost:5432/main
-spring.datasource.username=postgres
-spring.datasource.password=52458
+spring.datasource.url=jdbc:postgresql://localhost:5432/ewm-main
+spring.datasource.username=main
+spring.datasource.password=main
